Disable USB listDevice error check to fix Crazyradio on Windows
On Windows some devices cannot be open which results in an error returned by the
listDevice function. This can be safely ignored since what is important is if any Crazyadio
has been detected and can be used, not if other devices could not be open.

Also update the readme with instruction to install the Crazyradio driver on windows.